+++ alternator/CMakeLists.txt
@@ -9,6 +9,7 @@ target_sources(alternator
     controller.cc
     server.cc
     executor.cc
+    executor_read.cc
     stats.cc
     serialization.cc
     expressions.cc
